mgmt: mcumgr: Refactor callback header files
Refactors these so that each group defines the events it raises
