+++ include/zephyr/mgmt/mcumgr/mgmt/callbacks.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023 Nordic Semiconductor ASA
+ * Copyright (c) 2022-2026 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -10,24 +10,25 @@
 #include <inttypes.h>
 #include <zephyr/sys/slist.h>
 #include <zephyr/mgmt/mcumgr/mgmt/mgmt.h>
+#include "callback_defines.h"
 
-#ifdef CONFIG_MCUMGR_GRP_FS
+#if defined(CONFIG_MCUMGR_GRP_FS) || defined(__DOXYGEN__)
 #include <zephyr/mgmt/mcumgr/grp/fs_mgmt/fs_mgmt_callbacks.h>
 #endif
 
-#ifdef CONFIG_MCUMGR_GRP_IMG
+#if defined(CONFIG_MCUMGR_GRP_IMG) || defined(__DOXYGEN__)
 #include <zephyr/mgmt/mcumgr/grp/img_mgmt/img_mgmt_callbacks.h>
 #endif
 
-#ifdef CONFIG_MCUMGR_GRP_OS
+#if defined(CONFIG_MCUMGR_GRP_OS) || defined(__DOXYGEN__)
 #include <zephyr/mgmt/mcumgr/grp/os_mgmt/os_mgmt_callbacks.h>
 #endif
 
-#ifdef CONFIG_MCUMGR_GRP_SETTINGS
+#if defined(CONFIG_MCUMGR_GRP_SETTINGS) || defined(__DOXYGEN__)
 #include <zephyr/mgmt/mcumgr/grp/settings_mgmt/settings_mgmt_callbacks.h>
 #endif
 
-#ifdef CONFIG_MCUMGR_GRP_ENUM
+#if defined(CONFIG_MCUMGR_GRP_ENUM) || defined(__DOXYGEN__)
 #include <zephyr/mgmt/mcumgr/grp/enum_mgmt/enum_mgmt_callbacks.h>
 #endif
 
@@ -41,40 +42,6 @@ extern "C" {
  * @ingroup mcumgr
  * @{
  */
-
-/** @cond INTERNAL_HIDDEN */
-/** Event which signifies that all event IDs for a particular group should be enabled. */
-#define MGMT_EVT_OP_ID_ALL 0xffff
-
-/** Get event for a particular group and event ID. */
-#define MGMT_DEF_EVT_OP_ID(group, event_id) ((group << 16) | BIT(event_id))
-
-/** Get event used for enabling all event IDs of a particular group. */
-#define MGMT_DEF_EVT_OP_ALL(group) ((group << 16) | MGMT_EVT_OP_ID_ALL)
-/** @endcond  */
-
-/** Get group from event. */
-#define MGMT_EVT_GET_GROUP(event) ((event >> 16) & MGMT_EVT_OP_ID_ALL)
-
-/** Get event ID from event. */
-#define MGMT_EVT_GET_ID(event) (event & MGMT_EVT_OP_ID_ALL)
-
-/**
- * MGMT event callback return value.
- */
-enum mgmt_cb_return {
-	/** No error. */
-	MGMT_CB_OK,
-
-	/** SMP protocol error and ``err_rc`` contains the #mcumgr_err_t error code. */
-	MGMT_CB_ERROR_RC,
-
-	/**
-	 * Group (application-level) error and ``err_group`` contains the group ID that caused
-	 * the error and ``err_rc`` contains the error code of that group to return.
-	 */
-	MGMT_CB_ERROR_ERR,
-};
 
 /**
  * @typedef mgmt_cb
@@ -110,29 +77,6 @@ typedef enum mgmt_cb_return (*mgmt_cb)(uint32_t event, enum mgmt_cb_return prev_
 				       size_t data_size);
 
 /**
- * MGMT event callback group IDs. Note that this is not a 1:1 mapping with #mcumgr_group_t values.
- */
-enum mgmt_cb_groups {
-	MGMT_EVT_GRP_ALL			= 0,
-	MGMT_EVT_GRP_SMP,
-	MGMT_EVT_GRP_OS,
-	MGMT_EVT_GRP_IMG,
-	MGMT_EVT_GRP_FS,
-	MGMT_EVT_GRP_SETTINGS,
-	MGMT_EVT_GRP_ENUM,
-
-	MGMT_EVT_GRP_USER_CUSTOM_START		= MGMT_GROUP_ID_PERUSER,
-};
-
-/**
- * MGMT event opcodes for all command processing.
- */
-enum smp_all_events {
-	/** Used to enable all events. */
-	MGMT_EVT_OP_ALL				= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_ALL),
-};
-
-/**
  * MGMT event opcodes for base SMP command processing.
  */
 enum smp_group_events {
@@ -147,115 +91,6 @@ enum smp_group_events {
 
 	/** Used to enable all smp_group events. */
 	MGMT_EVT_OP_CMD_ALL			= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_SMP),
-};
-
-/**
- * MGMT event opcodes for filesystem management group.
- */
-enum fs_mgmt_group_events {
-	/** Callback when a file has been accessed, data is @ref fs_mgmt_file_access. */
-	MGMT_EVT_OP_FS_MGMT_FILE_ACCESS		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_FS, 0),
-
-	/** Callback when a file upload/download is finished, data is @ref fs_mgmt_file_access. */
-	MGMT_EVT_OP_FS_MGMT_FILE_ACCESS_DONE	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_FS, 1),
-
-	/** Used to enable all fs_mgmt_group events. */
-	MGMT_EVT_OP_FS_MGMT_ALL			= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_FS),
-};
-
-/**
- * MGMT event opcodes for image management group.
- */
-enum img_mgmt_group_events {
-	/** Callback when a client sends a file upload chunk, data is @ref img_mgmt_upload_check. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_CHUNK			= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 0),
-
-	/** Callback when a DFU operation is stopped. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_STOPPED		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 1),
-
-	/** Callback when a DFU operation is started. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_STARTED		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 2),
-
-	/** Callback when a DFU operation has finished being transferred. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_PENDING		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 3),
-
-	/** Callback when an image has been confirmed. data is @ref img_mgmt_image_confirmed. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_CONFIRMED		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 4),
-
-	/** Callback when an image write command has finished writing to flash. */
-	MGMT_EVT_OP_IMG_MGMT_DFU_CHUNK_WRITE_COMPLETE	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 5),
-
-	/**
-	 * Callback when an image slot's state is encoded for a response, data is
-	 * @ref img_mgmt_state_slot_encode.
-	 */
-	MGMT_EVT_OP_IMG_MGMT_IMAGE_SLOT_STATE		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 6),
-
-	/**
-	 * Callback when a slot list command outputs fields for an image, data is
-	 * @ref img_mgmt_slot_info_image.
-	 */
-	MGMT_EVT_OP_IMG_MGMT_SLOT_INFO_IMAGE		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 7),
-
-	/**
-	 * Callback when a slot list command outputs fields for a slot of an image, data is
-	 * @ref img_mgmt_slot_info_slot.
-	 */
-	MGMT_EVT_OP_IMG_MGMT_SLOT_INFO_SLOT		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_IMG, 8),
-
-	/** Used to enable all img_mgmt_group events. */
-	MGMT_EVT_OP_IMG_MGMT_ALL			= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_IMG),
-};
-
-/**
- * MGMT event opcodes for operating system management group.
- */
-enum os_mgmt_group_events {
-	/** Callback when a reset command has been received, data is @ref os_mgmt_reset_data. */
-	MGMT_EVT_OP_OS_MGMT_RESET		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 0),
-
-	/** Callback when an info command is processed, data is @ref os_mgmt_info_check. */
-	MGMT_EVT_OP_OS_MGMT_INFO_CHECK		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 1),
-
-	/** Callback when an info command needs to output data, data is @ref os_mgmt_info_append. */
-	MGMT_EVT_OP_OS_MGMT_INFO_APPEND		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 2),
-
-	/** Callback when a datetime get command has been received. */
-	MGMT_EVT_OP_OS_MGMT_DATETIME_GET	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 3),
-
-	/** Callback when a datetime set command has been received, data is @ref rtc_time. */
-	MGMT_EVT_OP_OS_MGMT_DATETIME_SET	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 4),
-
-	/**
-	 * Callback when a bootloader info command has been received, data is
-	 * os_mgmt_bootloader_info_data().
-	 */
-	MGMT_EVT_OP_OS_MGMT_BOOTLOADER_INFO	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_OS, 5),
-
-	/** Used to enable all os_mgmt_group events. */
-	MGMT_EVT_OP_OS_MGMT_ALL			= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_OS),
-};
-
-/**
- * MGMT event opcodes for settings management group.
- */
-enum settings_mgmt_group_events {
-	/** Callback when a setting is read/written/deleted. */
-	MGMT_EVT_OP_SETTINGS_MGMT_ACCESS	= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_SETTINGS, 0),
-
-	/** Used to enable all settings_mgmt_group events. */
-	MGMT_EVT_OP_SETTINGS_MGMT_ALL		= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_SETTINGS),
-};
-
-/**
- * MGMT event opcodes for enumeration management group.
- */
-enum enum_mgmt_group_events {
-	/** Callback when fetching details on supported command groups. */
-	MGMT_EVT_OP_ENUM_MGMT_DETAILS		= MGMT_DEF_EVT_OP_ID(MGMT_EVT_GRP_ENUM, 0),
-
-	/** Used to enable all enum_mgmt_group events. */
-	MGMT_EVT_OP_ENUM_MGMT_ALL		= MGMT_DEF_EVT_OP_ALL(MGMT_EVT_GRP_ENUM),
 };
 
 /**
